box64: 64-bit TLS/FLS -> UE4 FMallocBinned2 works -> Visage into file I/O
s_tls_slots/s_fls_slots: uint32_t -> uint64_t. TLS/FLS hold LPVOID guest pointers, and
FMallocBinned2 stashes its per-thread pool (a region-3 pointer >4GB) in TLS — a uint32
slot truncated it so the allocator read a low unmapped alias and crashed (r15=0x830000
= 0x10000830000 truncated). With 64-bit TLS the core UE4 allocator works; Visage now
reaches CreateFileW (config/asset I/O). 57 unique Win32 funcs.

Also: SIGSEGV handler now dumps R8/R9/R15/RBX/RDI + a guest .text stack trace (gated on
the crash) to pin truncations fast.

+++ Sources/Core/wg_engine.c
@@ -1021,9 +1021,12 @@ static struct { uint8_t used; uint32_t file_handle; } s_filemap[WG_MAX_FILEMAP];
 // reads 0 for an unset slot and the CRT lazily allocates its own block.
 // Indexed by scheduler thread slot (WG_MAX_THREADS rows). Windows guarantees
 // at least 1088 TLS slots.
-static uint32_t s_tls_slots[WG_MAX_THREADS][1088] = {{0}};
+// 64-bit slot values: TLS/FLS hold LPVOID (guest pointers). On the box64 path those
+// are >4GB (e.g. FMallocBinned2's per-thread pool at region-3), so a uint32 slot
+// truncated the pointer -> the allocator read a low unmapped alias and crashed.
+static uint64_t s_tls_slots[WG_MAX_THREADS][1088] = {{0}};
 static uint32_t s_tls_next = 0;
-static uint32_t s_fls_slots[WG_MAX_THREADS][1088] = {{0}};   // Fiber-Local Storage
+static uint64_t s_fls_slots[WG_MAX_THREADS][1088] = {{0}};   // Fiber-Local Storage
 static uint32_t s_fls_next = 0;
 
 // Fake event/mutex/semaphore handles. Single-threaded, so events are just
